Restructures documentation and commands
