+++ settings.gradle
@@ -4,3 +4,4 @@ pluginManagement {
 
 includeBuild '../formula'
 includeBuild '../feature-model'
+includeBuild '../formula-analysis-sat4j'
