+++ src/main/java/de/featjar/feature/model/io/uvl/UVLFormulaFormat.java
@@ -1,30 +1,77 @@
 package de.featjar.feature.model.io.uvl;
 
+import de.featjar.base.data.Problem;
 import de.featjar.base.data.Result;
 import de.featjar.base.io.format.IFormat;
 import de.featjar.base.io.input.AInputMapper;
+import de.featjar.base.tree.Trees;
+import de.featjar.feature.model.IFeatureModel;
+import de.featjar.feature.model.io.uvl.visitor.FormulaToUVLConstraintVisitor;
 import de.featjar.formula.analysis.combinations.BinomialCalculator;
 import de.featjar.formula.structure.formula.IFormula;
 import de.featjar.formula.structure.formula.connective.And;
 import de.featjar.formula.structure.formula.connective.Not;
 import de.featjar.formula.structure.formula.connective.Or;
+import de.featjar.formula.structure.formula.connective.Reference;
+import de.vill.main.UVLModelFactory;
+import de.vill.model.Attribute;
+import de.vill.model.Feature;
+import de.vill.model.Group;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class UVLFormulaFormat implements IFormat<IFormula> {
 
-    private BinomialCalculator bc;
-
-    public UVLFormulaFormat() {
-        bc = new BinomialCalculator(100,100);
-    }
+    public static final String ROOT_FEATURE_NAME = "Formula";
 
     @Override
     public Result<IFormula> parse(AInputMapper inputMapper) {
-        return Result.empty();
+        try {
+            String content = inputMapper.get().text();
+            UVLModelFactory uvlModelFactory = new UVLModelFactory();
+            de.vill.model.FeatureModel uvlModel = uvlModelFactory.parse(content);
+            IFeatureModel featureModel = UVLUtils.createFeatureModel(uvlModel);
+
+            List<IFormula> constraints = UVLUtils.uvlConstraintToFormula(uvlModel.getConstraints());
+
+            // TODO: Finish this
+
+            IFormula formula = new Reference(List.of());
+
+            return Result.of(formula);
+        } catch (Exception e) {
+            return Result.empty(e);
+        }
     }
 
     @Override
-    public Result<String> serialize(IFormula fm) {
-        return Result.empty();
+    public Result<String> serialize(IFormula formula) {
+        List<Problem> problems = new ArrayList<>();
+
+        de.vill.model.FeatureModel uvlModel = new de.vill.model.FeatureModel();
+        uvlModel.setNamespace("");
+        de.vill.model.Feature uvlRootFeature = new Feature(ROOT_FEATURE_NAME);
+        uvlRootFeature.getAttributes().put("abstract", new Attribute<>("abstract", true));
+        uvlModel.setRootFeature(uvlRootFeature);
+        uvlModel.getFeatureMap().put(ROOT_FEATURE_NAME, uvlRootFeature);
+
+        de.vill.model.Group uvlRootGroup = new Group(Group.GroupType.OR);
+        uvlRootFeature.addChildren(uvlRootGroup);
+
+        formula.getVariableNames().forEach((variableName) -> {
+            de.vill.model.Feature uvlFeature = new Feature(variableName);
+            uvlModel.getFeatureMap().put(variableName, uvlFeature);
+            uvlRootGroup.getFeatures().add(uvlFeature);
+        });
+
+        Result<de.vill.model.constraint.Constraint> uvlConstraint = Trees.traverse(formula, new FormulaToUVLConstraintVisitor());
+        problems.addAll(uvlConstraint.getProblems());
+        if (uvlConstraint.isEmpty()) {
+            return Result.empty(problems);
+        }
+        uvlModel.getConstraints().add(uvlConstraint.get());
+        return Result.of(uvlModel.toString(), problems);
     }
 
     @Override
@@ -45,68 +92,5 @@ public class UVLFormulaFormat implements IFormat<IFormula> {
     @Override
     public String getName() {
         return "Universal Variability Language";
-    }
-
-    protected IFormula[] nchoosek(IFormula[] elements, int k, boolean negated) {
-        final int n = elements.length;
-
-        // return tautology
-        if ((k == 0) || (k == (n + 1))) {
-            return new IFormula[] { new Or(new Not(elements[0]), elements[0])};
-        }
-
-        // return contradiction
-        if ((k < 0) || (k > (n + 1))) {
-            return new IFormula[] { new And(new Not(elements[0]), elements[0]) };
-        }
-
-        final IFormula[] newNodes = new IFormula[(int) binomial(n, k)];
-        int j = 0;
-
-        // negate all elements
-        if (negated) {
-            negateNodes(elements);
-        }
-
-        final IFormula[] clause = new IFormula[k];
-        final int[] index = new int[k];
-
-        // the position that is currently filled in clause
-        int level = 0;
-        index[level] = -1;
-
-        while (level >= 0) {
-            // fill this level with the next element
-            index[level]++;
-            // did we reach the maximum for this level
-            if (index[level] >= (n - (k - 1 - level))) {
-                // go to previous level
-                level--;
-            } else {
-                clause[level] = elements[index[level]];
-                if (level == (k - 1)) {
-                    newNodes[j++] = new Or(clause);
-                } else {
-                    // go to next level
-                    level++;
-                    // allow only ascending orders (to prevent from duplicates)
-                    index[level] = index[level - 1];
-                }
-            }
-        }
-        if (j != newNodes.length) {
-            throw new RuntimeException("Pre-calculation of the number of elements failed!");
-        }
-        return newNodes;
-    }
-
-    public static long binomial(int n, int k) {
-       return new BinomialCalculator(n, k).binomial();
-    }
-
-    protected static void negateNodes(IFormula[] nodes) {
-        for (int i = 0; i < nodes.length; i++) {
-            nodes[i] = new Not(nodes[i]);
-        }
     }
 }
