+++ src/main/resources/extensions.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0"?>
 <extensions>
     <point id="de.featjar.feature.model.io.FeatureModelFormats">
-        <extension id="de.featjar.feature.model.io.xml.UVLFeatureModelFormat" />
+        <extension id="de.featjar.feature.model.io.uvl.UVLFeatureModelFormat" />
+        <extension id="de.featjar.feature.model.io.uvl.UVLFormulaFormat" />
     </point>
 </extensions>
